requirements.txt
pygame-ce
opencv-python
ffpyplayer
numpy